# docs/swagger.yaml
basePath: /
definitions:
  models.ProductDescription:
    properties:
      additional_info:
        type: string
      color:
        type: string
      gender:
        type: string
      id:
        type: integer
      material:
        type: string
    type: object
  request.AddToCartRequest:
    properties:
      productId:
        type: integer
      size:
        type: string
    type: object
  request.CustomerLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  request.CustomerRegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  request.ProductRequest:
    properties:
      productId:
        type: integer
    type: object
  request.ProductsByCategoryRequest:
    properties:
      categoryName:
        type: string
    type: object
  response.Category:
    properties:
      name:
        type: string
    type: object
  response.PopProductsResponse:
    properties:
      products:
        items:
          $ref: '#/definitions/response.ProductResponsePart'
        type: array
    type: object
  response.ProductCategoryResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/response.Category'
        type: array
    type: object
  response.ProductResponse:
    properties:
      description:
        $ref: '#/definitions/models.ProductDescription'
      image_url:
        type: string
      name:
        type: string
      price:
        type: string
      productId:
        type: integer
      sizes:
        items:
          type: string
        type: array
    type: object
  response.ProductResponsePart:
    properties:
      description_info:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: string
      productId:
        type: integer
    type: object
  response.ProductsByCategoryResponse:
    properties:
      products:
        items:
          $ref: '#/definitions/response.ProductResponsePart'
        type: array
    type: object
  response.TokenResponse:
    properties:
      token:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: This is a Bebeziansback
  title: Bebeziansback
  version: "1.0"
paths:
  /api/:
    get:
      description: Responds with the test answer as JSON.
      responses: {}
      summary: Get Test
      tags:
      - Server
  /api/cart/addToCart:
    post:
      consumes:
      - application/json
      description: Add Product To Cart
      parameters:
      - description: Cart Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Add To Cart
      tags:
      - Cart
  /api/customer/register:
    post:
      consumes:
      - application/json
      description: Register Customer with name and password
      parameters:
      - description: Customer
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CustomerRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Customer created successfully
      summary: Register Customer
      tags:
      - Customer
  /api/product/getAllCategory:
    get:
      description: Get all product categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductCategoryResponse'
      summary: Get All Categories
      tags:
      - Product
  /api/product/getPopProducts:
    get:
      description: Get all Pop Products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PopProductsResponse'
      summary: Get All Pop Products
      tags:
      - Product
  /api/product/getProduct:
    post:
      consumes:
      - application/json
      description: Get Product by id
      parameters:
      - description: Product Id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductResponse'
      summary: Get Product
      tags:
      - Product
  /api/product/getProductsByCategory:
    post:
      consumes:
      - application/json
      description: Get products by category name
      parameters:
      - description: Category Name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.ProductsByCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductsByCategoryResponse'
      summary: Get All Products By Category
      tags:
      - Product
  /api/profile/info:
    get:
      consumes:
      - application/json
      description: Get Profile Info by customer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Get Profile Info
      tags:
      - Profile
  /cart/getCart:
    get:
      consumes:
      - application/json
      description: Get Products from Cart by customer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Get Cart
      tags:
      - Cart
  /customer/login:
    post:
      consumes:
      - application/json
      description: Login Customer with name and password
      parameters:
      - description: Customer
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CustomerLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Customer logged in successfully
          schema:
            $ref: '#/definitions/response.TokenResponse'
      summary: Login Customer
      tags:
      - Customer
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"